Add initial integration with dome billing
