+++ src/settings.py
@@ -273,6 +273,8 @@ INVENTORY = environ.get("BAE_CB_INVENTORY", INVENTORY)
 RESOURCE_INVENTORY = environ.get("BAE_CB_RESOURCE_INVENTORY", RESOURCE_INVENTORY)
 SERVICE_INVENTORY = environ.get("BAE_CB_SERVICE_INVENTORY", SERVICE_INVENTORY)
 
+DOME_BILLING_URL = environ.get("BAE_CB_DOME_BILLING_URL", DOME_BILLING_URL)
+
 ORDERING = environ.get("BAE_CB_ORDERING", ORDERING)
 BILLING = environ.get("BAE_CB_BILLING", BILLING)
 USAGE = environ.get("BAE_CB_USAGE", USAGE)
@@ -295,3 +297,6 @@ if isinstance(ACL_ENABLED, str):
 AWS_ENABLED = environ.get("AWS_ENABLED", False)
 if isinstance(AWS_ENABLED, str):
     AWS_ENABLED = AWS_ENABLED == "True"
+
+
+BILLING_ENGINE = 'dome'
